% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seo_helper_functions.R
\name{base_points_calculate}
\alias{base_points_calculate}
\title{base_points_calculate}
\usage{
base_points_calculate(X, value, bA1)
}
\arguments{
\item{X}{\itemize{
\item 
}}

\item{value}{\itemize{
\item 
}}

\item{bA1}{\itemize{
\item 
}}
}
\description{
calculate base woz points
}
