% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seo_helper_functions.R
\name{calc.verhuurdersheffing}
\alias{calc.verhuurdersheffing}
\title{calc.verhuurdersheffing}
\usage{
calc.verhuurdersheffing(df, percentage, aftopping)
}
\arguments{
\item{df}{\itemize{
\item 
}}

\item{percentage}{\itemize{
\item 
}}

\item{aftopping}{\itemize{
\item 
}}
}
\description{
calc.verhuurdersheffing
}
