% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seo_helper_functions.R
\name{calc_huurstijging}
\alias{calc_huurstijging}
\title{calc_huurstijging}
\usage{
calc_huurstijging(scenario_name, df_list_vec)
}
\arguments{
\item{scenario_name}{bla}

\item{df_list_vec}{bla}
}
\value{
bla
}
\description{
bla
}
