% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_box3.R
\name{calculate_box3}
\alias{calculate_box3}
\title{Calculate Box3  tax , based,  assets and debt. Uses leegwaarde and vrijstellingsfactor to come to correct}
\usage{
calculate_box3(
  houses,
  debt_share,
  rent_func,
  transfer_tax = 0.104,
  lwr = "nieuw",
  box3_percentage_2025onwards = 0.34,
  investor = "part_box3",
  forfait_schijf2_oud = 0.0467,
  box3_oud = "nee",
  vermogen = 1e+05
)
}
\arguments{
\item{houses}{Houses df}

\item{debt_share}{Percentage of home value as mortgage debt}

\item{lwr}{'nieuw' or 'oud' for old or new lwr system}

\item{investor}{Investor type string}

\item{forfait_schijf2_oud}{Assumed return in 2022 box 3}

\item{box3_oud}{'nieuw' or 'oud' for 2023 and after or 2022 system}

\item{vermogen}{Wealth outside of property}
}
\value{
A set of columns used to determine pre-post tax cashflows
}
\description{
Calculate Box3  tax , based,  assets and debt. Uses leegwaarde and vrijstellingsfactor to come to correct
}
