% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_debt.R
\name{calculate_debt}
\alias{calculate_debt}
\title{Calculate Debt of the investment}
\usage{
calculate_debt(houses, debt_share)
}
\arguments{
\item{houses}{Houses df}

\item{debt_share}{Percentage of home value as mortgage debt}

\item{transfer_tax}{Transfer tax rate}
}
\value{
A column with debt based on inputs purchase price, debt-share and transfer tax
}
\description{
Calculate Debt of the investment
}
