% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_energy_investment.R
\name{calculate_energy_investment}
\alias{calculate_energy_investment}
\title{Calculate Verduurzaming. If label is under a certain target label, investor will invest more money to improve energy label. This is a function of m2, a reference table with cost and a target label}
\usage{
calculate_energy_investment(
  houses,
  energy_investment_matrix,
  energylabelgoalval = 4
)
}
\arguments{
\item{houses}{Houses df.}

\item{energy_investment_matrix}{Matrix with costs per m2 for every step from one label to another.}

\item{energylabelgoalval}{The minimum energy label goal.}
}
\value{
two column with the cost to get to the target energy label. cost per m2 and total cost.
}
\description{
Calculate Verduurzaming. If label is under a certain target label, investor will invest more money to improve energy label. This is a function of m2, a reference table with cost and a target label
}
