% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_woz.R
\name{calculate_house_price_index}
\alias{calculate_house_price_index}
\title{Functions for estimating WOZ value in 2022}
\usage{
calculate_house_price_index(houseprice_index, houses)
}
\arguments{
\item{houses}{Houses df}
}
\value{
A vector with as many elements as houses has rows.
}
\description{
Functions for estimating WOZ value in 2022
}
