% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_irr.R
\name{calculate_irr}
\alias{calculate_irr}
\title{Calculate IRR}
\usage{
calculate_irr(cash_flows)
}
\arguments{
\item{cash_flows}{Vector of cash flows}
}
\value{
A IRR value
}
\description{
Calculate IRR
}
