% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_market_rent.R
\name{calculate_market_rent}
\alias{calculate_market_rent}
\title{Calculate market rent}
\usage{
calculate_market_rent(houses, market_rent_table)
}
\arguments{
\item{houses}{Houses df}

\item{market_rent_table}{}
}
\value{
a vector of marketrent per year
}
\description{
Calculate market rent
}
