% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_npv.R
\name{calculate_npv}
\alias{calculate_npv}
\title{Calculate NPV}
\usage{
calculate_npv(cash_flows, discount_rate)
}
\arguments{
\item{cash_flows}{Vector of cashflows}

\item{discount_rate}{Discount rate (i.e. 0.05)}
}
\value{
A NPV value
}
\description{
Calculate NPV
}
