% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rent.R
\name{calculate_rent_all_houses}
\alias{calculate_rent_all_houses}
\title{Calculate rent for all housing sector}
\usage{
calculate_rent_all_houses(houses, growth_newcomer, mutation_grade, parameters)
}
\arguments{
\item{houses}{Houses df}

\item{growth_newcomer}{Rent increase when occupant changes}

\item{mutation_grade}{Mutation grade of occupants}

\item{growth_stayer}{Rent increase with same occupant}
}
\description{
Calculate rent for all housing sector
}
