% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_returns.R
\name{calculate_returns}
\alias{calculate_returns}
\title{Calculate returns}
\usage{
calculate_returns(houses, discount_rate, ...)
}
\arguments{
\item{houses}{Houses df}

\item{discount_rate}{Discount rate (i.e. 0.05)}

\item{...}{Arbitrary number of functions that take houses as an argument and return a vector of cashflows}
}
\value{
A list object with multiple types of returns
}
\description{
Calculate returns
}
