% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transfer_tax.R
\name{calculate_transfer_tax}
\alias{calculate_transfer_tax}
\title{Calculate Debt of the investment}
\usage{
calculate_transfer_tax(houses, transfer_tax)
}
\arguments{
\item{houses}{Houses df}

\item{transfer_tax}{Transfer tax rate}
}
\value{
A column with a negative transfer tax cashflow, only in the first year
}
\description{
Calculate Debt of the investment
}
