% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_vpb.R
\name{calculate_vpb_and_box2_tax}
\alias{calculate_vpb_and_box2_tax}
\title{Calculate Box 2  tax , based,  assets and debt.
Source VPB rate: https://www.belastingdienst.nl/wps/wcm/connect/bldcontentnl/belastingdienst/zakelijk/winst/vennootschapsbelasting/tarieven_vennootschapsbelasting
Source dividend rate: https://www.rijksoverheid.nl/onderwerpen/inkomstenbelasting/2-belastingschijven-box-2}
\usage{
calculate_vpb_and_box2_tax(
  houses,
  returns_func,
  vpb_rate = 0.26,
  box2_rate = 0.245
)
}
\arguments{
\item{houses}{houses}

\item{returns_func}{Functions with all relevant cashflows (including sale)}

\item{vpb_rate}{VPB tax rate}

\item{box2_rate}{Dividend tax rate}
}
\description{
Calculate Box 2  tax , based,  assets and debt.
Source VPB rate: https://www.belastingdienst.nl/wps/wcm/connect/bldcontentnl/belastingdienst/zakelijk/winst/vennootschapsbelasting/tarieven_vennootschapsbelasting
Source dividend rate: https://www.rijksoverheid.nl/onderwerpen/inkomstenbelasting/2-belastingschijven-box-2
}
