% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_wws_points.R
\name{calculate_wws_points}
\alias{calculate_wws_points}
\title{calculate_wws_points}
\usage{
calculate_wws_points(houses, wws_reform, bA1 = 1, bA91 = 11041, bA91b = 172)
}
\arguments{
\item{houses}{\itemize{
\item 
}}

\item{bA1}{\itemize{
\item 
}}

\item{bA91}{\itemize{
\item 
}}

\item{bA91b}{\itemize{
\item 
}}

\item{delta_value}{\itemize{
\item 
}}
}
\description{
calculate_wws_points
}
