% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seo_helper_functions.R
\name{calculate_wws_points_old}
\alias{calculate_wws_points_old}
\title{calculate_wws_points}
\usage{
calculate_wws_points_old(
  df_list_vec,
  delta_value,
  bA1,
  bA91,
  bA91b,
  lookup.table
)
}
\arguments{
\item{df_list_vec}{\itemize{
\item 
}}

\item{delta_value}{\itemize{
\item 
}}

\item{bA1}{\itemize{
\item 
}}

\item{bA91}{\itemize{
\item 
}}

\item{bA91b}{\itemize{
\item 
}}

\item{lookup.table}{\itemize{
\item 
}}
}
\description{
calculate_wws_points
}
