% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seo_helper_functions.R
\name{calculate_wws_rent}
\alias{calculate_wws_rent}
\title{calculate_wws_rent}
\usage{
calculate_wws_rent(houses, lookup.table)
}
\arguments{
\item{houses}{\itemize{
\item 
}}

\item{lookup.table}{\itemize{
\item 
}}
}
\description{
Get huur.wws based on wws.point
}
