% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seo_helper_functions.R
\name{call.box3_oud}
\alias{call.box3_oud}
\title{call.box3_oud}
\usage{
call.box3_oud(
  vermogen,
  heffingsvrij,
  schijf1,
  schijf2,
  forf1,
  forf2,
  forf3,
  perc3
)
}
\arguments{
\item{vermogen}{\itemize{
\item 
}}

\item{heffingsvrij}{\itemize{
\item 
}}

\item{schijf1}{\itemize{
\item 
}}

\item{schijf2}{\itemize{
\item 
}}

\item{forf1}{\itemize{
\item 
}}

\item{forf2}{\itemize{
\item 
}}

\item{forf3}{\itemize{
\item 
}}

\item{perc3}{\itemize{
\item 
}}
}
\description{
call.box3_oud
}
