% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seo_helper_functions.R
\name{call.eindwaarde.doorexploiteren}
\alias{call.eindwaarde.doorexploiteren}
\title{call.eindwaarde.doorexploiteren}
\usage{
call.eindwaarde.doorexploiteren(CF, d, g, m)
}
\arguments{
\item{CF}{bla}

\item{d}{bla}

\item{g}{bla}

\item{m}{bla}
}
\value{
bla
}
\description{
bla
}
