% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seo_helper_functions.R
\name{call.leegwaarde}
\alias{call.leegwaarde}
\title{call.leegwaarde}
\usage{
call.leegwaarde(huur, WOZ)
}
\arguments{
\item{huur}{\itemize{
\item 
}}

\item{WOZ}{\itemize{
\item 
}}
}
\description{
call.leegwaarde
}
