% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seo_helper_functions.R
\name{call.plot.huis}
\alias{call.plot.huis}
\title{call.plot.huis}
\usage{
call.plot.huis(df, huisnr, vars)
}
\arguments{
\item{df}{bla}

\item{huisnr}{bla}

\item{vars}{bla}
}
\value{
bla
}
\description{
call.plot.huis
}
