% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_rent_at_return.R
\name{cap_rent_at_return}
\alias{cap_rent_at_return}
\title{cap_rent_at_return}
\usage{
cap_rent_at_return(houses, rent_func, rent_categories, return_perc)
}
\arguments{
\item{houses}{the generic houses dataframe}

\item{rent_func}{A function that only needs houses as an argument to produce a cashflow stream}

\item{rent_categories}{Rent categories to which the cap should be applied}

\item{cap_perc}{The height of the rent cap}
}
\value{
Single column with cashflows
}
\description{
cap_rent_at_return
}
