% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_rent_at_woz_perc.R
\name{cap_rent_at_woz_perc}
\alias{cap_rent_at_woz_perc}
\title{cap_rent_at_woz_perc}
\usage{
cap_rent_at_woz_perc(
  houses,
  rent_func,
  rent_categories,
  cap_perc,
  interest_rate = 0,
  interest_rate_weight = 0
)
}
\arguments{
\item{houses}{the generic houses dataframe}

\item{rent_func}{A function that only needs houses as an argument to produce a cashflow stream}

\item{rent_categories}{Rent categories to which the cap should be applied}

\item{cap_perc}{The height of the rent cap}
}
\value{
Single column with cashflows
}
\description{
cap_rent_at_woz_perc
}
