% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_rent_direct_return.R
\name{cap_rent_direct_return}
\alias{cap_rent_direct_return}
\title{cap_rent_direct_return}
\usage{
cap_rent_direct_return(
  houses,
  house_costs_func,
  rent_func,
  rent_categories,
  direct_return_cap,
  debt_share
)
}
\arguments{
\item{houses}{the generic houses dataframe}

\item{house_costs_func}{A partialised get_returns function}

\item{rent_func}{A function that only needs houses as an argument to produce a cashflow stream}

\item{rent_categories}{Rent categories to which the cap should be applied}

\item{cap_perc}{The height of the rent cap}
}
\value{
Single column with cashflows
}
\description{
cap_rent_direct_return
}
