% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rent_index_scenario.R
\name{construct_indices}
\alias{construct_indices}
\title{Get max rent increase index, based on a specific scenario that corresponds with
#'}
\usage{
construct_indices(scenario_name)
}
\arguments{
\item{scenario_name}{string of a scenario that corresponds with an}
}
\value{
A IRR value
}
\description{
Get max rent increase index, based on a specific scenario that corresponds with
#'
}
