% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_wws_with_woz_cap.R
\name{correct_woz_cap}
\alias{correct_woz_cap}
\title{calculate capped points if condition is met. Overwrite the woz_points with corrected amount of points
calc_capped_points}
\usage{
correct_woz_cap(houses, threshold = 142, bA93i = 0.33)
}
\arguments{
\item{threshold}{}
}
\description{
calculate capped points if condition is met. Overwrite the woz_points with corrected amount of points
calc_capped_points
}
