% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rent.R
\name{create_rent_category}
\alias{create_rent_category}
\title{Rent category}
\usage{
create_rent_category(houses, points_social_boundary, points_middle_boundary)
}
\arguments{
\item{houses}{Houses df}

\item{points_social_boundary}{Point up to which social housing}

\item{points_middle_boundary}{Point up to which 'middenhuur'}
}
\description{
Assign a rent sector a property is in in a given year.
}
