% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_wws_points.R
\name{.base_points_calculate}
\alias{.base_points_calculate}
\title{Calculate on surface and energylabel points.}
\usage{
.base_points_calculate(houses, wws_reform, bA1)
}
\arguments{
\item{houses}{\itemize{
\item 
}}

\item{wws_reform}{\itemize{
\item 
}}

\item{bA1}{\itemize{
\item 
}}
}
\description{
Calculate on surface and energylabel points.
}
