% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_wws_points.R
\name{.points_correct_egwmgw}
\alias{.points_correct_egwmgw}
\title{points_correct_egwmgew}
\usage{
.points_correct_egwmgw(houses, extra.points.egw, extra.points.mgw)
}
\arguments{
\item{houses}{\itemize{
\item 
}}
}
\description{
points_correct_egwmgew
}
