% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seo_beleidsvarianten.R
\name{get_beleidsparameters}
\alias{get_beleidsparameters}
\title{SEO beleidsvarianten vector}
\usage{
get_beleidsparameters(beleid, variant_num)
}
\arguments{
\item{beleid}{Beleid df (read in pipeline)}

\item{variant_num}{The policy variant index}
}
\value{
bla
}
\description{
bla
}
