% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uniform_rent_price.R
\name{get_uniform_rent_price}
\alias{get_uniform_rent_price}
\title{Functions for creating uniform shell rental ask price}
\usage{
get_uniform_rent_price(
  houses,
  correction_upholstered = 0.94,
  correction_furnished = 0.85
)
}
\arguments{
\item{houses}{Houses df}

\item{correction_upholstered}{Percentage correction to define uniform house rental. 0.90 -> 95\% of rental price is for shell.}
}
\value{
A vector with as many elements as houses has rows.
}
\description{
Functions for creating uniform shell rental ask price
}
