% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_purchase_price.R
\name{index_purchase_price}
\alias{index_purchase_price}
\title{Functions to index house purhcase price.}
\usage{
index_purchase_price(houses)
}
\arguments{
\item{houses}{Houses df}
}
\value{
A vector with as many elements as houses has rows.
}
\description{
Functions to index house purhcase price.
}
