% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest.R
\name{interest_constant_interest_rate}
\alias{interest_constant_interest_rate}
\title{Functions for interest payments}
\usage{
interest_constant_interest_rate(houses, debt_share, interest_rate)
}
\arguments{
\item{houses}{Houses df}

\item{debt_share}{What percentage of the house market value}

\item{interest_rate}{Interest rate on mortgage debt (eg 0.05 for a 5\% interest rate)}
}
\value{
A vector with as many elements as houses has rows.
}
\description{
Functions for interest payments
}
