% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investment.R
\name{investment_constant_upkeep}
\alias{investment_constant_upkeep}
\title{Functions for investments and upkeep}
\usage{
investment_constant_upkeep(houses, upkeep_perc)
}
\arguments{
\item{houses}{Houses df}

\item{upkeep_perc}{Percentage of home value spent on upkeep yearly (eg 0.01)}
}
\value{
A vector with as many elements as houses has rows.
}
\description{
Functions for investments and upkeep
}
