% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investment.R
\name{investment_exploitation_cost}
\alias{investment_exploitation_cost}
\title{Exploitation cost}
\usage{
investment_exploitation_cost(houses, rent_func, exploitation_cost_perc)
}
\arguments{
\item{houses}{Houses df}

\item{exploitation_cost}{Percentage of rent spent on exploitation  (eg 0.2)}
}
\value{
A vector with as many elements as houses has rows.
}
\description{
Exploitation cost
}
