% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investment.R
\name{investment_incidental_cost}
\alias{investment_incidental_cost}
\title{Incidental cost}
\usage{
investment_incidental_cost(houses, incidental_cost, investment_year)
}
\arguments{
\item{houses}{Houses df}

\item{incidental_cost}{number of init cost of rent spent on exploitation  (eg 0.2)}

\item{investment_year}{Modeled year of investment}
}
\value{
A vector with as many elements as houses has rows.
}
\description{
Incidental cost
}
