% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_all_numeric}
\alias{is_all_numeric}
\title{Helper file with all simple functions -> columns to numeric where possible}
\usage{
is_all_numeric(x)
}
\arguments{
\item{x}{A data.frame}
}
\value{
df
}
\description{
Helper file with all simple functions -> columns to numeric where possible
}
