% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nl_grid_returns.R
\name{make_nl_grid_returns}
\alias{make_nl_grid_returns}
\title{Make Geo Grid object with returns}
\usage{
make_nl_grid_returns(returns_by_postal_code, num_hex)
}
\arguments{
\item{returns_by_postal_code}{Returns by postal code}

\item{num_hex}{Number of cells in grid}
}
\value{
Geo object
}
\description{
Make Geo Grid object with returns
}
