% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middenhuur_bump.R
\name{middenhuur_rent_bump}
\alias{middenhuur_rent_bump}
\title{Calculate returns given a new policy}
\usage{
middenhuur_rent_bump(houses, middenhuur_bump_percentage)
}
\arguments{
\item{houses}{the generic houses dataframe}

\item{middenhuur_bump_percentage}{percentage to increase wws rent}
}
\value{
houses df with newly calculated wws_rent given the percentage. Only applied to middenhuur houses.
}
\description{
Calculate returns given a new policy
}
