% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_highest_cashflow.R
\name{pick_highest_cashflow}
\alias{pick_highest_cashflow}
\title{Pick Highest Cashflow}
\usage{
pick_highest_cashflow(houses, discount_rate, ..., name = "highest_cashflow")
}
\arguments{
\item{houses}{Houses df}

\item{discount_rate}{Single value such as 0.058}

\item{...}{Arbitrary number of cashflow generating functions}

\item{name}{Optional name for the resulting column}
}
\value{
A single cashflow
}
\description{
For each unique id and policy variant, one of multiple mutually exclusive cash flow options is chosen.
The chosen option is the one that has the highest total discounted value.
The discounting is perfomed with the provided discount value.
}
