% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_sale_or_rent.R
\name{pick_sale_or_rent}
\alias{pick_sale_or_rent}
\title{Register in Houses the Values of Selling and Renting, and indicate which one is preferable}
\usage{
pick_sale_or_rent(houses, parameters, funcs)
}
\arguments{
\item{houses}{Houses df}

\item{keep_rent_func}{}
}
\value{
A single cashflow
}
\description{
Register in Houses the Values of Selling and Renting, and indicate which one is preferable
}
