% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seo_helper_functions.R
\name{plot_huur_list}
\alias{plot_huur_list}
\title{plot_huur_list}
\usage{
plot_huur_list(house_num, var_list)
}
\arguments{
\item{house_num}{\itemize{
\item 
}}

\item{var_list}{\itemize{
\item 
}}
}
\description{
plot_huur_list
}
