% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seo_helper_functions.R
\name{read_huur_verhoging_scenario}
\alias{read_huur_verhoging_scenario}
\title{read_huur_verhoging_scenario}
\usage{
read_huur_verhoging_scenario(scenario_name)
}
\arguments{
\item{scenario_name}{\itemize{
\item 
}}
}
\value{
index
}
\description{
read_huur_verhoging_scenario
}
