% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rent.R
\name{rent_constant}
\alias{rent_constant}
\title{Functions for rent}
\usage{
rent_constant(houses)
}
\arguments{
\item{houses}{Houses df}
}
\value{
A vector with as many elements as houses has rows.
}
\description{
Functions for rent
}
