% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rent.R
\name{rent_constant_increase}
\alias{rent_constant_increase}
\title{rent_constant_increase}
\usage{
rent_constant_increase(houses, increase_factor = 0.02)
}
\arguments{
\item{houses}{Houses df}

\item{increase_factor}{Percentage increase per year (ie 0.02)}
}
\description{
rent_constant_increase
}
