% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_box3_scenario.R
\name{run_box3_scenario}
\alias{run_box3_scenario}
\title{Run Standard Scenario}
\usage{
run_box3_scenario(houses, parameters)

run_box3_scenario(houses, parameters)
}
\arguments{
\item{houses}{Houses df}

\item{upkeep_perc}{Upkeep cost as percentage of property value}

\item{rent_increase}{Yearly rent increase (i.e. 0.02)}

\item{debt_share}{The share of the property value financed with debt}

\item{interest_rate}{Constant interest rate.}

\item{discount_rate}{Discount rate (i.e. 0.05)}

\item{exploitation_cost}{\% of rent for upkeep etc}

\item{incidental_investment_year}{Modeled year of investment}

\item{incidental_cost}{Cost at incidental_investment_year}

\item{refinance_period}{After how many years the investor refinances the property}
}
\value{
A list object with multiple types of returns

A list object with multiple types of returns
}
\description{
A scenario consists of multiple cashflow streams. The 'standard' scenario
includes rent, sale/further rental at the end of the investment period,
upkeep costs and taxes. This function makes it easy to run the standard
scenario.

A scenario consists of multiple cashflow streams. The 'standard' scenario
includes rent, sale/further rental at the end of the investment period,
upkeep costs and taxes. This function makes it easy to run the standard
scenario.
}
