% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scenario.R
\name{run_scenario}
\alias{run_scenario}
\title{Run  Scenario}
\usage{
run_scenario(houses, parameters, funcs)
}
\arguments{
\item{houses}{Houses df}

\item{funcs}{a list with the elements rent_func, cap_func, aq_cost_func, running_cost_funcs, sale_func, keep_rent_func, and tax_func}
}
\value{
A list object with multiple types of returns
}
\description{
Run  Scenario
}
