% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seo_helper_functions.R
\name{setup.data.frame_vec}
\alias{setup.data.frame_vec}
\title{SEO helper functions}
\usage{
setup.data.frame_vec(first.year, n.year, huisaantal, variantaantal)
}
\arguments{
\item{first.year}{First simulation year}

\item{n.year}{Number of simulation years}

\item{huisaantal}{Number of simulation houses}

\item{variantaantal}{Number of policy variants}
}
\value{
A vector with as many elements as houses has rows.
}
\description{
SEO helper functions
}
