% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{transaction_buy_at_market_value}
\alias{transaction_buy_at_market_value}
\title{Functions for property transactions}
\usage{
transaction_buy_at_market_value(houses, parameters, funcs)
}
\arguments{
\item{houses}{Houses df}
}
\value{
A vector with as many elements as houses has rows.
}
\description{
Functions for property transactions
}
