% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{transaction_keep_property_for_rent}
\alias{transaction_keep_property_for_rent}
\title{Keep property for Rent}
\usage{
transaction_keep_property_for_rent(
  houses,
  rent_cashflow,
  rent_increase = 0.02,
  discount_rate = 0.0513
)
}
\arguments{
\item{houses}{Houses df}

\item{rent_cashflow}{Rental income}

\item{rent_increase}{Assumed yearly rent increase (0.01 = 1\% assumed rent increase)}

\item{discount_rate}{Investor discount rate for future returns (such as 0.058)}
}
\description{
Keep property for Rent
}
