% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{transaction_sell_at_orignal_value}
\alias{transaction_sell_at_orignal_value}
\title{transaction_sell_at_orignal_value}
\usage{
transaction_sell_at_orignal_value(houses)
}
\arguments{
\item{houses}{Houses df}
}
\description{
transaction_sell_at_orignal_value
}
