% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{transaction_sell_value_constant_increase}
\alias{transaction_sell_value_constant_increase}
\title{transaction_sell_value_constant_increase}
\usage{
transaction_sell_value_constant_increase(houses, parameters = parameters)
}
\arguments{
\item{houses}{Houses df}

\item{increase_factor}{yearly price increase}
}
\description{
transaction_sell_value_constant_increase
}
