% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seo_helper_functions.R
\name{verduurzamen}
\alias{verduurzamen}
\title{verduurzamen}
\usage{
verduurzamen(m2, start_label, label_target)
}
\arguments{
\item{m2}{\itemize{
\item 
}}

\item{start_label}{\itemize{
\item 
}}

\item{label_target}{\itemize{
\item 
}}
}
\description{
verduurzamen
}
