% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seo_helper_functions.R
\name{verhoging}
\alias{verhoging}
\title{verhoging}
\usage{
verhoging(
  huur_tmin1,
  pntgrens.sociaal,
  pntgrens.midden,
  wws.punten,
  max.groei.sociaal,
  max.groei.sociaal.part,
  max.groei.midden,
  max.groei.vrij,
  huur.wws,
  vrije.huur,
  jaar,
  type,
  opslag_middenhuur,
  huur_oud,
  max.groei.sociaal.oud,
  i
)
}
\arguments{
\item{huur_tmin1}{\itemize{
\item 
}}

\item{pntgrens.sociaal}{\itemize{
\item 
}}

\item{pntgrens.midden}{\itemize{
\item 
}}

\item{wws.punten}{\itemize{
\item 
}}

\item{max.groei.sociaal}{\itemize{
\item 
}}

\item{max.groei.sociaal.part}{\itemize{
\item 
}}

\item{max.groei.midden}{\itemize{
\item 
}}

\item{max.groei.vrij}{\itemize{
\item 
}}

\item{huur.wws}{\itemize{
\item 
}}

\item{vrije.huur}{\itemize{
\item 
}}

\item{jaar}{\itemize{
\item 
}}

\item{type}{\itemize{
\item 
}}

\item{opslag_middenhuur}{\itemize{
\item 
}}

\item{huur_oud}{\itemize{
\item 
}}

\item{max.groei.sociaal.oud}{\itemize{
\item 
}}

\item{i}{bla}
}
\value{
bla
}
\description{
verhoging
}
