% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seo_helper_functions.R
\name{wws.helper}
\alias{wws.helper}
\title{Title}
\usage{
wws.helper(
  m2,
  energy.label,
  finish.level,
  woz,
  points.a1,
  points.a4,
  points.a56,
  points.a9,
  lookup.table,
  woz_index,
  type,
  bA1,
  bA91,
  bA91b
)
}
\arguments{
\item{m2}{\itemize{
\item 
}}

\item{energy.label}{bla}

\item{finish.level}{bla}

\item{woz}{bla}

\item{points.a1}{bla}

\item{points.a4}{bla}

\item{points.a56}{bla}

\item{points.a9}{bla}

\item{lookup.table}{bla}

\item{woz_index}{bla}

\item{type}{bla}

\item{bA1}{bla}

\item{bA91}{bla}

\item{bA91b}{bla}
}
\value{
bla
}
\description{
Title
}
